feat(day-five): navigation
Sorted out what I wanted to do for the navigation. After going back and
forth between how to manage each day as a component, now I have a better
idea of how to make the pages work. Unfortunately, now the code is
repeating itself again. A refactor would be nice to make it easier to
folllow.