Dev (#8)
* fix(day-one): remove last set recovery

Timer fired after last set was complete. No need for that.

* feat(day-five): add selection buttons

* feat: workouts are components

self-contained components. Still playing around with the idea of having
the workouts be components instead of pages. However, I am feeling some
UX issue here with navigation.

* feat(day-five): navigation

Sorted out what I wanted to do for the navigation. After going back and
forth between how to manage each day as a component, now I have a better
idea of how to make the pages work. Unfortunately, now the code is
repeating itself again. A refactor would be nice to make it easier to
folllow.

* fix(console): remove console.log

* feat(day-5): update buttons and wording

* feat(fonts): use different fonts

* feat(emoji): add save emoji

Also, adjust styling on day input to match other days' inputs

* fix(data): collect workout data

daily workout data is collected when the save emoji is hit. Now I need
to work on sending it to indexedDB

* refactor: rename an abbreviation

* feat(action buttons): resize

* feat: style program page

* feat: style program page, again

* fix(indexedDB): start implementing indexedDB

Getting more comfortable with indexedDB, but it is mental
gymnastics. Haha

* refactor: indexedDB and components

moved around a bunch of files and updated indexedDB to update week data
when a workout is completed

* feat(indexedDB): implement indexedDB

Initialize and use indexedDB to control which day is shown to users when
they begin the program.